Added second set of latches
Added code for second set of latches that will be written to but not displayed until the 4th clock cycle. the code written for the RPI will need to be sure and understand this. It's always writing to the next line to be displayed, Not the current line being displayed.